moved column explanation to helper file, for better visibility
